(maint) get expand working properly with /events
This is a nasty workaround that patches over a code issue, which is that the
events endpoint passes a vector [query-options paging-options] where the other
endpoints pass a map paging-options. We could fix this by passing an options
map or by combining query-options and paging-options on the events endpoint.

We should fix the underlying issue before this code is merged; this is just to
help get testing working on the cdb-retirements branch for now.